Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int a,b,c,average;
+    float a,b,c,average;
     scanf("%d %d %d",&a,&b,&c);
     average=(a+b+c)/3;
     printf("Averge: %.2f",average);
